Rename user service and move to auth package
